# GODRIVER-3533 Optimize value reader and writer