--- a/bson/value_writer.go
+++ b/bson/value_writer.go
@@ -35,7 +35,7 @@ func putValueWriter(vw *valueWriter) {
 
 var documentWriterPool = sync.Pool{
 	New: func() interface{} {
-		return NewDocumentWriter(nil)
+		return newDocumentWriter(nil)
 	},
 }