… bson/mem.prof