<!--- If applicable, issue number goes here, e.g. GODRIVER-ABCD -->
GODRIVER-3533

## Summary

Use sync.Pool for the value writer and reader when encoding and decoding, respectively. Additionally, use peek and discard when appending elements to the read dst. **The danger here occurs when a user finds a way to call Get() more than Put()** which would result in a "leak" of reusable objects. 

Note: Regressions from v1 will be addressed in GODRIVER-3450. The primary goal of this PR is to add optimizations that were removed in the v2 implementation.

### Marshaling Benchmarks 

v1: 

```
BenchmarkMarshal/BSON/simple_struct-10           5279845               227.6 ns/op           288 B/op          1 allocs/op
BenchmarkMarshal/BSON/nested_struct-10           2844418               440.9 ns/op           416 B/op          1 allocs/op
BenchmarkMarshal/BSON/simple_D-10                5591744               217.8 ns/op           288 B/op          1 allocs/op
BenchmarkMarshal/BSON/deep_bson.json.gz-10        188263              6385 ns/op            9114 B/op        316 allocs/op
BenchmarkMarshal/BSON/flat_bson.json.gz-10        166884              7429 ns/op           14901 B/op        292 allocs/op
BenchmarkMarshal/BSON/full_bson.json.gz-10        168651              7226 ns/op           11001 B/op        241 allocs/op
BenchmarkMarshal/extJSON/simple_struct-10         695290              1683 ns/op            3477 B/op         88 allocs/op
BenchmarkMarshal/extJSON/nested_struct-10         510933              2423 ns/op            4630 B/op        132 allocs/op
BenchmarkMarshal/extJSON/simple_D-10              712291              1664 ns/op            3476 B/op         88 allocs/op
BenchmarkMarshal/extJSON/deep_bson.json.gz-10      73616             16458 ns/op           30228 B/op        890 allocs/op
BenchmarkMarshal/extJSON/flat_bson.json.gz-10      33312             36623 ns/op           83594 B/op       1284 allocs/op
BenchmarkMarshal/extJSON/full_bson.json.gz-10      36608             32847 ns/op           68587 B/op       1232 allocs/op
BenchmarkMarshal/JSON/simple_struct-10           8541211               144.9 ns/op           240 B/op          1 allocs/op
BenchmarkMarshal/JSON/nested_struct-10           6256928               183.3 ns/op           352 B/op          1 allocs/op
BenchmarkMarshal/JSON/simple_D-10                4428652               281.7 ns/op           448 B/op          1 allocs/op
BenchmarkMarshal/JSON/deep_bson.json.gz-10        200554              6352 ns/op           10869 B/op        316 allocs/op
BenchmarkMarshal/JSON/flat_bson.json.gz-10        103041             11485 ns/op           17848 B/op        295 allocs/op
BenchmarkMarshal/JSON/full_bson.json.gz-10        130994              9218 ns/op           14130 B/op        235 allocs/op
PASS
ok      go.mongodb.org/mongo-driver/bson        25.208s
```

PR:

```
BenchmarkMarshal/BSON/simple_struct-10           5395154               216.7 ns/op           288 B/op          1 allocs/op
BenchmarkMarshal/BSON/nested_struct-10           2889548               445.5 ns/op           416 B/op          1 allocs/op
BenchmarkMarshal/BSON/simple_D-10                5884207               210.3 ns/op           288 B/op          1 allocs/op
BenchmarkMarshal/BSON/deep_bson.json.gz-10        592386              1990 ns/op            2162 B/op          6 allocs/op
BenchmarkMarshal/BSON/flat_bson.json.gz-10        185100              6838 ns/op           14898 B/op        292 allocs/op
BenchmarkMarshal/BSON/full_bson.json.gz-10        194402              6245 ns/op           10332 B/op        211 allocs/op
BenchmarkMarshal/extJSON/simple_struct-10         761506              1609 ns/op            3476 B/op         88 allocs/op
BenchmarkMarshal/extJSON/nested_struct-10         512126              2270 ns/op            4630 B/op        132 allocs/op
BenchmarkMarshal/extJSON/simple_D-10              712059              1649 ns/op            3477 B/op         88 allocs/op
BenchmarkMarshal/extJSON/deep_bson.json.gz-10     112938             10419 ns/op           23267 B/op        580 allocs/op
BenchmarkMarshal/extJSON/flat_bson.json.gz-10      34946             34078 ns/op           83585 B/op       1284 allocs/op
BenchmarkMarshal/extJSON/full_bson.json.gz-10      12500             92065 ns/op          188265 B/op       1369 allocs/op
BenchmarkMarshal/JSON/simple_struct-10           8987240               134.8 ns/op           240 B/op          1 allocs/op
BenchmarkMarshal/JSON/nested_struct-10           7025131               182.9 ns/op           352 B/op          1 allocs/op
BenchmarkMarshal/JSON/simple_D-10                1390707               775.8 ns/op           929 B/op         17 allocs/op
BenchmarkMarshal/JSON/deep_bson.json.gz-10         77088             15683 ns/op           21436 B/op        298 allocs/op
BenchmarkMarshal/JSON/flat_bson.json.gz-10        120090             10125 ns/op           17616 B/op        293 allocs/op
BenchmarkMarshal/JSON/full_bson.json.gz-10        120982              9982 ns/op           16129 B/op        270 allocs/op
PASS
ok      go.mongodb.org/mongo-driver/v2/bson     26.065s
``` 

Note that we improved in deep BSON test cases but regressed with full BSON test cases. Otherwise the results are fairly similar. 

### Unmarshaling Benchmarks 

v1:
```
BenchmarkUnmarshal/BSON/simple_struct-10                 3398468               347.9 ns/op       773.22 MB/s         640 B/op         27 allocs/op
BenchmarkUnmarshal/BSON/nested_struct-10                 8039444               139.4 ns/op      2955.96 MB/s         328 B/op          4 allocs/op
BenchmarkUnmarshal/BSON/simple_to_map-10                 1000000              1154 ns/op         233.10 MB/s        2051 B/op         66 allocs/op
BenchmarkUnmarshal/BSON/simple_to_D-10                   1529571               779.5 ns/op       345.11 MB/s        1576 B/op         34 allocs/op
BenchmarkUnmarshal/BSON/nested_to_map-10                  331148              3646 ns/op         113.01 MB/s        8461 B/op        146 allocs/op
BenchmarkUnmarshal/BSON/nested_to_D-10                    548484              2205 ns/op         186.84 MB/s        4818 B/op         92 allocs/op
BenchmarkUnmarshal/BSON/deep_bson.json.gz_to_map-10                81966             14420 ns/op         136.34 MB/s       31172 B/op        825 allocs/op
BenchmarkUnmarshal/BSON/deep_bson.json.gz_to_D-10                 144742              8436 ns/op         233.06 MB/s       14093 B/op        573 allocs/op
BenchmarkUnmarshal/BSON/flat_bson.json.gz_to_map-10                66225             17945 ns/op         336.92 MB/s       33596 B/op        753 allocs/op
BenchmarkUnmarshal/BSON/flat_bson.json.gz_to_D-10                 165055              7163 ns/op         844.00 MB/s       15853 B/op        316 allocs/op
BenchmarkUnmarshal/BSON/full_bson.json.gz_to_map-10                67675             18299 ns/op         231.27 MB/s       30682 B/op        808 allocs/op
BenchmarkUnmarshal/BSON/full_bson.json.gz_to_D-10                  98814             12325 ns/op         343.38 MB/s       23794 B/op        511 allocs/op
BenchmarkUnmarshal/extJSON/simple_struct-10                       349196              3441 ns/op         111.59 MB/s        6931 B/op        201 allocs/op
BenchmarkUnmarshal/extJSON/nested_struct-10                       273610              4361 ns/op         115.81 MB/s        9204 B/op        271 allocs/op
BenchmarkUnmarshal/extJSON/simple_to_map-10                       280131              4320 ns/op          88.88 MB/s        8342 B/op        240 allocs/op
BenchmarkUnmarshal/extJSON/simple_to_D-10                         347761              3970 ns/op          96.74 MB/s        7868 B/op        208 allocs/op
BenchmarkUnmarshal/extJSON/nested_to_map-10                       147412              8172 ns/op          61.80 MB/s       17074 B/op        387 allocs/op
BenchmarkUnmarshal/extJSON/nested_to_D-10                         177060              6717 ns/op          75.18 MB/s       13429 B/op        333 allocs/op
BenchmarkUnmarshal/extJSON/deep_bson.json.gz_to_map-10             36777             33408 ns/op          53.10 MB/s       64852 B/op       1779 allocs/op
BenchmarkUnmarshal/extJSON/deep_bson.json.gz_to_D-10               46696             26018 ns/op          68.18 MB/s       47692 B/op       1527 allocs/op
BenchmarkUnmarshal/extJSON/flat_bson.json.gz_to_map-10             21060             56535 ns/op         144.23 MB/s      101612 B/op       2673 allocs/op
BenchmarkUnmarshal/extJSON/flat_bson.json.gz_to_D-10               25857             45972 ns/op         177.37 MB/s       83853 B/op       2235 allocs/op
BenchmarkUnmarshal/extJSON/full_bson.json.gz_to_map-10             15014             76837 ns/op          88.95 MB/s      120958 B/op       3488 allocs/op
BenchmarkUnmarshal/extJSON/full_bson.json.gz_to_D-10               16902             70240 ns/op          97.31 MB/s      114074 B/op       3191 allocs/op
BenchmarkUnmarshal/JSON/simple_struct-10                         2839286               425.9 ns/op       540.08 MB/s         360 B/op          9 allocs/op
BenchmarkUnmarshal/JSON/nested_struct-10                         2434580               498.2 ns/op       704.55 MB/s         520 B/op          7 allocs/op
BenchmarkUnmarshal/JSON/simple_to_map-10                         1215229               996.1 ns/op       230.90 MB/s        1682 B/op         50 allocs/op
--- FAIL: BenchmarkUnmarshal/JSON/simple_to_D
    benchmark_test.go:427: error unmarshalling JSON: json: cannot unmarshal object into Go value of type primitive.D
BenchmarkUnmarshal/JSON/nested_to_map-10                          512389              2257 ns/op         155.49 MB/s        5538 B/op         76 allocs/op
--- FAIL: BenchmarkUnmarshal/JSON/nested_to_D
    benchmark_test.go:427: error unmarshalling JSON: json: cannot unmarshal object into Go value of type primitive.D
BenchmarkUnmarshal/JSON/deep_bson.json.gz_to_map-10               113818             10240 ns/op         173.25 MB/s       23864 B/op        389 allocs/op
--- FAIL: BenchmarkUnmarshal/JSON/deep_bson.json.gz_to_D
    benchmark_test.go:427: error unmarshalling JSON: json: cannot unmarshal object into Go value of type primitive.D
BenchmarkUnmarshal/JSON/flat_bson.json.gz_to_map-10                63711             19287 ns/op         351.07 MB/s       30372 B/op        548 allocs/op
--- FAIL: BenchmarkUnmarshal/JSON/flat_bson.json.gz_to_D
    benchmark_test.go:427: error unmarshalling JSON: json: cannot unmarshal object into Go value of type primitive.D
BenchmarkUnmarshal/JSON/full_bson.json.gz_to_map-10                51853             22884 ns/op         232.22 MB/s       40788 B/op        764 allocs/op
--- FAIL: BenchmarkUnmarshal/JSON/full_bson.json.gz_to_D
    benchmark_test.go:427: error unmarshalling JSON: json: cannot unmarshal object into Go value of type primitive.D
--- FAIL: BenchmarkUnmarshal/JSON
--- FAIL: BenchmarkUnmarshal
FAIL
exit status 1
FAIL    go.mongodb.org/mongo-driver/bson        46.022s
```

PR:
```
BenchmarkUnmarshal/BSON/simple_struct-10                 2687252               435.8 ns/op       617.22 MB/s         594 B/op         43 allocs/op
BenchmarkUnmarshal/BSON/nested_struct-10                 5361310               223.5 ns/op      1843.79 MB/s         609 B/op          5 allocs/op
BenchmarkUnmarshal/BSON/simple_to_map-10                 1000000              1098 ns/op         244.94 MB/s        2006 B/op         82 allocs/op
BenchmarkUnmarshal/BSON/simple_to_D-10                   1298235               916.2 ns/op       293.61 MB/s        1723 B/op         62 allocs/op
BenchmarkUnmarshal/BSON/nested_to_map-10                  570775              1950 ns/op         211.28 MB/s        3494 B/op        140 allocs/op
BenchmarkUnmarshal/BSON/nested_to_D-10                    635196              1864 ns/op         221.03 MB/s        3221 B/op        139 allocs/op
BenchmarkUnmarshal/BSON/deep_bson.json.gz_to_map-10               119232             10405 ns/op         188.94 MB/s       16993 B/op        889 allocs/op
BenchmarkUnmarshal/BSON/deep_bson.json.gz_to_D-10                 116686             10105 ns/op         194.56 MB/s       16751 B/op        887 allocs/op
BenchmarkUnmarshal/BSON/flat_bson.json.gz_to_map-10                67744             19374 ns/op         312.07 MB/s       39974 B/op        923 allocs/op
BenchmarkUnmarshal/BSON/flat_bson.json.gz_to_D-10                 108649             11504 ns/op         525.57 MB/s       24534 B/op        630 allocs/op
BenchmarkUnmarshal/BSON/full_bson.json.gz_to_map-10                63975             19073 ns/op         221.88 MB/s       33227 B/op       1042 allocs/op
BenchmarkUnmarshal/BSON/full_bson.json.gz_to_D-10                  77899             15481 ns/op         273.38 MB/s       29404 B/op        860 allocs/op
BenchmarkUnmarshal/extJSON/simple_struct-10                       355491              3324 ns/op         115.51 MB/s        6931 B/op        201 allocs/op
BenchmarkUnmarshal/extJSON/nested_struct-10                       272911              4301 ns/op         117.42 MB/s        9204 B/op        271 allocs/op
BenchmarkUnmarshal/extJSON/simple_to_map-10                       283300              4178 ns/op          91.91 MB/s        8341 B/op        240 allocs/op
BenchmarkUnmarshal/extJSON/simple_to_D-10                         300330              3906 ns/op          98.30 MB/s        8059 B/op        220 allocs/op
BenchmarkUnmarshal/extJSON/nested_to_map-10                       173290              6956 ns/op          72.60 MB/s       14070 B/op        357 allocs/op
BenchmarkUnmarshal/extJSON/nested_to_D-10                         173780              6552 ns/op          77.07 MB/s       13797 B/op        356 allocs/op
BenchmarkUnmarshal/extJSON/deep_bson.json.gz_to_map-10             44523             26567 ns/op          66.77 MB/s       49980 B/op       1655 allocs/op
BenchmarkUnmarshal/extJSON/deep_bson.json.gz_to_D-10               47103             26318 ns/op          67.41 MB/s       49739 B/op       1653 allocs/op
BenchmarkUnmarshal/extJSON/flat_bson.json.gz_to_map-10             22389             53443 ns/op         152.57 MB/s      101592 B/op       2673 allocs/op
BenchmarkUnmarshal/extJSON/flat_bson.json.gz_to_D-10               27526             42824 ns/op         190.41 MB/s       86181 B/op       2380 allocs/op
BenchmarkUnmarshal/extJSON/full_bson.json.gz_to_map-10             17000             68351 ns/op         100.00 MB/s      119552 B/op       3476 allocs/op
BenchmarkUnmarshal/extJSON/full_bson.json.gz_to_D-10               17534             67082 ns/op         101.89 MB/s      115747 B/op       3294 allocs/op
BenchmarkUnmarshal/JSON/simple_struct-10                         2837935               405.6 ns/op       567.00 MB/s         360 B/op          9 allocs/op
BenchmarkUnmarshal/JSON/nested_struct-10                         2484396               474.1 ns/op       740.35 MB/s         520 B/op          7 allocs/op
BenchmarkUnmarshal/JSON/simple_to_map-10                         1255458               919.8 ns/op       250.07 MB/s        1682 B/op         50 allocs/op
BenchmarkUnmarshal/JSON/simple_to_D-10                            419559              2815 ns/op          81.69 MB/s        5127 B/op        199 allocs/op
BenchmarkUnmarshal/JSON/nested_to_map-10                          492206              2309 ns/op         152.02 MB/s        5538 B/op         76 allocs/op
BenchmarkUnmarshal/JSON/nested_to_D-10                            276355              4400 ns/op          79.77 MB/s        7624 B/op        317 allocs/op
BenchmarkUnmarshal/JSON/deep_bson.json.gz_to_map-10               115366             10117 ns/op         175.35 MB/s       23928 B/op        389 allocs/op
BenchmarkUnmarshal/JSON/deep_bson.json.gz_to_D-10                  56762             21424 ns/op          82.81 MB/s       35191 B/op       1722 allocs/op
BenchmarkUnmarshal/JSON/flat_bson.json.gz_to_map-10            64996         18060 ns/op         374.92 MB/s       30367 B/op        548 allocs/op
BenchmarkUnmarshal/JSON/flat_bson.json.gz_to_D-10              31407         37441 ns/op         180.85 MB/s       53649 B/op       2288 allocs/op
BenchmarkUnmarshal/JSON/full_bson.json.gz_to_map-10                65449             18243 ns/op         267.61 MB/s       32170 B/op        616 allocs/op
BenchmarkUnmarshal/JSON/full_bson.json.gz_to_D-10                  23976             41944 ns/op         116.39 MB/s       64659 B/op       2766 allocs/op
PASS
ok      go.mongodb.org/mongo-driver/v2/bson     52.558s
```

## Background & Motivation

Some optimizations were removed while flatting the bson library during the v2 implementation. Specifically:

- https://github.com/mongodb/mongo-go-driver/pull/1732
- https://github.com/mongodb/mongo-go-driver/pull/1410